use a static archive instead of a framework
