--- Cartfile
@@ -1,1 +1,0 @@
-binary "https://storage.googleapis.com/fiskaly-cdn/clients/com.fiskaly.client-ios-all-v1.1.400.carthage.json" == 1.1.400
