+++ FiskalySDK/Classes/FiskalyHttpClient.swift
@@ -1,12 +1,4 @@
 import Foundation
-import FiskalyClient
-
-func FiskalyClientInvoke(_ request: String) -> String {
-    let resultRaw = _fiskaly_client_invoke(request)
-    let result = String(cString: resultRaw!)
-    _fiskaly_client_free(resultRaw)
-    return result
-}
 
 public class FiskalyHttpClient {
