+++ FiskalySDK/Classes/JsonRpc/JsonRpcError.swift
