+++ FiskalySDK/Classes/JsonRpc/JsonRpcRequest.swift
