+++ FiskalySDK/Classes/JsonRpc/JsonRpcResponse.swift
