+++ FiskalySDK/Classes/Results.swift
