+++ FiskalySDK/FiskalyClient.h
@@ -1,0 +1,4 @@
+
+extern const char* _fiskaly_client_invoke(const char* request);
+
+extern void _fiskaly_client_free(const char* response);
